Fixes for OpenMP < 3
Apparently in OpenMP versions below 3, loop indices must be
signed integers, and unsigned integers don't work. This
makes FFLD2 fail to compile on Windows, so silence the
warnings by some ugly static casting.

+++ HOGPyramid.cpp
@@ -122,7 +122,7 @@ void HOGPyramid::convolve(const Level & filter, vector<Matrix> & convolutions) c
 	convolutions.resize(levels_.size());
 	
 #pragma omp parallel for
-	for (unsigned int i = 0; i < levels_.size(); ++i)
+	for (int i = 0; i < static_cast<int>(levels_.size()); ++i)
 		Convolve(levels_[i], filter, convolutions[i]);
 }
